api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
google-generativeai>=0.4.0 # Add Gemini library
python-dotenv # Good practice for managing API keys via .env file (optional but recommended)

